move table row container to own file
